:iphone: Refactorizar SignIn
